# backend/docs/swagger.yaml
basePath: /api
definitions:
  models.DifficultyLevel:
    enum:
    - beginner
    - intermediate
    - advanced
    type: string
    x-enum-varnames:
    - Beginner
    - Intermediate
    - Advanced
  models.Employment:
    enum:
    - frontend
    - backend
    - fullstack
    - mobile
    type: string
    x-enum-varnames:
    - EmploymentFrontend
    - EmploymentBackend
    - EmploymentFullstack
    - EmploymentMobile
  models.Experience:
    enum:
    - basic
    - intermediate
    - advanced
    type: string
    x-enum-varnames:
    - BasicLevel
    - IntermediateLevel
    - AdvancedLevel
  models.Level:
    enum:
    - junior
    - pleno
    - senior
    type: string
    x-enum-varnames:
    - JuniorLevel
    - PlenoLevel
    - SeniorLevel
...
  schemas.CreateProjectRequest:
    properties:
      backend_techs:
        items:
          type: string
        type: array
      delivery_date:
        type: string
      description:
        type: string
      estimated_time:
        type: integer
      frontend_techs:
        items:
          type: string
        type: array
      level:
        type: string
      quanty_max_users:
        type: integer
      title:
        type: string
    required:
    - delivery_date
    - description
    - estimated_time
    - quanty_max_users
    - title
    type: object
  schemas.CreateRoadmapRequest:
    properties:
      area:
        type: string
      content:
        type: string
      course_links:
        type: string
      difficulty:
        $ref: '#/definitions/models.DifficultyLevel'
      title:
        type: string
    required:
    - area
    - content
    - difficulty
    - title
    type: object
...
  schemas.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      employment:
        allOf:
        - $ref: '#/definitions/models.Employment'
        example: backend
      experience:
        allOf:
        - $ref: '#/definitions/models.Experience'
        example: basic
      level:
        allOf:
        - $ref: '#/definitions/models.Level'
        example: junior
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - employment
    - experience
    - level
    - password
    - username
    type: object
  schemas.UserIDRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
host: localhost:5000
info:
  contact:
    email: support@hublast.com
    name: API Support
  description: Project Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: HubLast-Hub API
  version: "1.0"
paths:
  /admin/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project by its ID (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates an existing project with the provided details
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated project details
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns message and updated project
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
...
  /admin/roadmaps/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing roadmap (admin only)
      parameters:
      - description: Roadmap ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Roadmap deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete roadmap
      tags:
      - roadmaps
    put:
      consumes:
      - application/json
      description: Update an existing roadmap (admin only)
      parameters:
      - description: Roadmap ID
        in: path
        name: id
        required: true
        type: integer
      - description: Roadmap information
        in: body
        name: roadmap
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateRoadmapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Roadmap updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Roadmap not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update roadmap information
      tags:
      - roadmaps
...
  /api/admin/projects/{id}/responsible:
    post:
      consumes:
      - application/json
      description: Sets a user as responsible for a project (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to set as responsible
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.UserIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project responsible set successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set project responsible
      tags:
      - projects
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"